Move the E0130 check to AST validation pass
